Workaround for "specialized" > "constrained"
C++20's partially ordered overloads prefer specialized overloads before
constrained overloads. This means that we cannot simply use a variadic
lambda. Instead, the approach is now to overload lambdas for each
possible number of arguments.

Noexcept friendliness is unsupported.
Variadic parameters are unsupported.
Captures are unsupported, because we don't have a way to tell what
should be used to capture captures inside the overloaded lambdas. This
may be reasonable anyway, because captures behave irregularly with
`decltype(auto)`.
Empty argument lists are unsupported, because requires expressions are
ill-formed if it's not possible for their expressions to be valid.
These can be added in later. Variadics require creating a pack `_args`
for each overload, which should be possible with the nesting of another
lambda. Inside the requires expression, this may require lambdas in
unevaluated contexts.